pyproject.toml
[project]
name = "nebulapicker"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Silva",email = "djsilva@gmx.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = ".", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
coverage = "^7.10.6"
ruff = "^0.12.11"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Pytest ini configuration
[tool.pytest.ini_options]
pythonpath = ". src"
addopts = "--pyargs"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
omit = ["*/tests/*"]
branch = true

[tool.coverage.report]
fail_under = 85
show_missing = true

# Ruff configuration
[tool.ruff]
src = ["app"]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "UP", # pyupgrade
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "PT", # flake8-pytest-style
    "N", # pep8-naming
    "C90", # mccabe
]
line-length = 99
target-version = "py311"
